backend/requirements.txt
fastapi
uvicorn
python-dotenv
pymupdf
groq
python-multipart
